Added security

+++ src/main/dbscripts/incremental/01_v1.0/01_CreateLoanTable.sql
@@ -1,7 +1,7 @@
 create table Loan(
 	LoanID BIGINT PRIMARY KEY AUTO_INCREMENT,
 	StartingPrincipal INTEGER NOT NULL,
-	StartingInterest FLOAT NOT NULL,
+	StartingInterest DECIMAL(20,6) NOT NULL,
 	StartingFees INTEGER NOT NULL,
-	Margin FLOAT NOT NULL
+	Margin DECIMAL(20,6) NOT NULL
 );
